refactor single-resource api calls into a singular endpoint
